Implement storage upon opendal Operator
